Added api to support saving the currently opend db to a backup file; Refactored all io calls
