+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "onekeepass-core"
-version = "0.10.0"
+version = "0.11.0"
 authors = ["jeyasankar"]
 edition = "2021"
